% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Audi}
\alias{Audi}
\title{Audi data}
\format{
Sales prices and technical data on 30 cars of type Audi A4.
The data were collected in February 2017. The variables are: Price, Km, Hk,
Transition, Volume, Fuel, CO2, Weight, year, Age where Price is in 1000 NOK,Km
is the distance driven (in 1000 km),Hk the horse power,Transition either manual
(M) or automatic (A) transmission, etc.
}
\usage{
Audi
}
\description{
Data for Group-Exercises-Regression I
}
\examples{
summary(Audi)
}
\keyword{datasets}
