% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{barley}
\alias{barley}
\title{barley data}
\format{
In barley.rdata are results from an experiment where the response
is yield of barley pr 1000 square meter, and the factors sorts of barley,
soil types, types of fertilizers. In addition,the experiment was done in
two different geographical areas (sites).
}
\usage{
barley
}
\description{
A dataset with 32 observations
}
\examples{
summary(barley)
}
\keyword{datasets}
