% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Birds}
\alias{Birds}
\title{Birds data}
\format{
A data frame with 25 rows (one for each bird species) and 3163 columns.
The value reflects the largest simultaneous count
of a given bird species in a square region .
}
\usage{
Birds
}
\description{
Birds were observed in 10 km2 squares across entire mainland Norway,
which was divided into 3163 squares in total.
Every square is considered one variable. We
consider the 25 most abundant winter bird species in Norway.
}
\seealso{
\code{Birds4}
}
\keyword{datasets}
