% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Birds4.R
\docType{data}
\name{Birds4}
\alias{Birds4}
\title{Birds4: Winter Bird Species in Norway (subset)}
\format{
A data frame with 4 observations (rows) and 3163 variables (columns).

Every square is considered one variable. The value reflects the largest
simultaneous count of a given bird species in a square region.
}
\usage{
Birds4
}
\description{
This data set is a subset of the \code{Birds} data. Birds were observed in
10 km\eqn{^2} squares across entire mainland Norway, which was divided into
3163 squares in total.
}
\details{
We consider 4 winter bird species in Norway:
\itemize{
\item common redpoll (gråsisik)
\item fieldfare (gråtrost)
\item mallard (stokkand)
\item crow (kråke)
}
}
\examples{

# First five columns
Birds4[, 1:5]

}
\seealso{
\code{Birds} and \code{Birdcount}
}
\keyword{datasets}
