% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{birth}
\alias{birth}
\title{birth data}
\format{
A data frame with 4 four continuous variables
\itemize{
\item \code{LOW} NO or YES (low birth weight)
\item \code{AGE} in years
\item \code{LWT} mothers weight in pound before pregnancy
\item \code{SMK} NO or YES (smoker)
\item \code{BTW} birth weight in g
}
}
\usage{
birth
}
\description{
The dataset records of 189 birth weights
from Massachusetts, USA, and some additional variables.
}
\examples{
lm(BWT ~ LWT*SMK, data = birth)
}
\keyword{datasets}
