% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boilers.R
\docType{data}
\name{boilers}
\alias{boilers}
\title{boilers: Erecting Boiler Drums}
\format{
A data frame with 36 observations (rows) and 5 variables (columns).

\tabular{llllr}{
\tab \bold{Column name} \tab \bold{Data type} \tab \bold{Description} \tab \bold{Values}\cr
\code{[,1]} \tab \code{Hours} \tab integer \tab Man-hours \tab (1200 - 14791)\cr
\code{[,2]} \tab \code{Capacity} \tab integer \tab Boiler capacity in lb/hour \tab (30000 - 1089490)\cr
\code{[,3]} \tab \code{Pressure} \tab integer \tab Boiler design pressure in PSI \tab (250 - 2170)\cr
\code{[,4]} \tab \code{Boiler} \tab integer \tab 1 if industry erect, otherwise 0 \tab (0 - 1)\cr
\code{[,5]} \tab \code{Drum} \tab integer \tab Steam drum = 1, mud drum = 0 \tab (0 - 1)
}
}
\source{
Mendenhall, W., Sincich, T. (2012) \emph{A second course in statistics:
Regression Analysis, 7th edition.} Pearson Education.
}
\usage{
boilers
}
\description{
This data contains man-hours required to erect boiler drums of two types:
steam and mud drums. A manufacturer of boiler drums wants to use regression
to predict the number of hours needed to erect the drums in future projects.
To accomplish this, information of 35 boilers were collected; the boiler
design pressure in pounds per square inch (PSI), boiler capacity and if the
boiler was industry field erected or utility field erected.
}
\details{
A boiler is used to provide hot water or steam for heating. The steam drum is
located in the upper drum of a boiler and used to store the steam generated
in water-tube boilers. A mud drum, which is the lower drum, can be added to
the system to collect and remove sediments and mud in the water.

This data is from Exercise 4.74 in A second course in statistics: Regression
Analysis, 7th edition.
}
\examples{

# First eight rows
head(boilers, n = 8)

# Man-hours for industry erected mud boiler drums
idx <- which(boilers$Boiler == 1 & boilers$Drum == 0)
boxplot(boilers$Hours[idx] ~ boilers$Pressure[idx])

}
\keyword{datasets}
