% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bummary.R
\name{bummary}
\alias{bummary}
\title{Data summary}
\usage{
bummary(data)
}
\arguments{
\item{data}{A data table (data.frame or tibble).}
}
\value{
A data.frame with one row for each column in the argument \code{data}
and with the columns:
\itemize{
\item variable. Column name.
\item type. The data type.
\item cases. The number of observations.
\item missing. The number of missing data.
\item minimum. Minimum value (numeric and integer only).
\item quantile_0.25. First quartile (numeric and integer only).
\item median. Median value (numeric and integer only).
\item quantile_0.75. Third quartile (numeric and integer only).
\item maximum. Maximum value (numeric and integer only).
\item mean. Mean value (numeric and integer only).
\item sd. Standard deviation (numeric and integer only).
\item n_categories. The number of categories (factor and logical only).
\item levels. A list-column with factor level counts (factor and logical only).
}
}
\description{
Improved summary function.
}
\details{
This is a small function for making better summary of data,
especially for big data sets (many variables/columns). The output is given
as a data.frame with one row for each column in the argument \code{data}, thus
we tolerate many columns. Also, the summary output includes the number of
missing data for each variable and its data type. For numerical variables
the basic summary statistics are computed, and for factors the count for
each factor level is given, in the order of the factor levels.

The argument \code{data} is expected
to have columns of data types \code{character}, \code{integer}, \code{numeric}, \code{factor} or
\code{logical}. Columns with \code{list} are tolerated, but will not be summarized.
}
\examples{
library(BIAS.data)
data(colon)
summary.tbl <- bummary(colon)

}
\author{
Lars Snipen.
}
