% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Cakes.miss}
\alias{Cakes.miss}
\title{Cakes.miss data}
\format{
there are 5 cakes left to bake (response Y set to NA in observation
nr. 9, 11, 13, 15, and 20), and you are supposed to complete the data set
using the virtual chocolate cake factory.
}
\usage{
Cakes.miss
}
\description{
Data for Exercise 4 - "Finding the best chocolate cake recipe" in Group-Exercises-RegressionII
}
\examples{
summary(Cakes.miss)
}
\keyword{datasets}
