% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{colon}
\alias{colon}
\title{colon data}
\format{
The data set is a 62 × 2001 matrix.
Each column corresponds to the expression level of a certain gene.
The the rows correspond to different tissues.
The higher the value, the more active the gene is in a given tissue.
The first column Tissue codes the disease status.
}
\usage{
colon
}
\description{
Data from a colon cancer study (Alon et al. (1999) Proc. Natl. Acad. Sci. U.S.A. 96 (12): 6745–6750).
}
\examples{
summary(colon)
}
\keyword{datasets}
