% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{crabs}
\alias{crabs}
\title{crabs data}
\format{
There are 173 female crabs for which we wish to model
the presence or absence of male satellites dependent upon characteristics
of the female horseshoe crabs. Here, 𝑦=1if satellite is present and 0
otherwise. Explanatory variables are: weight, width of shell, color
(medium light, medium, medium dark, dark), and condition of spine.
}
\usage{
crabs
}
\description{
A dataset with 173 observations
}
\examples{
summary(crabs)
}
\keyword{datasets}
