% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D.R
\docType{data}
\name{D}
\alias{D}
\title{D: Distance Matrix}
\format{
A data frame with 4 rows and 4 columns.
}
\usage{
D
}
\description{
This data set is a 4 by 4 distance matrix. It can be used to illustrate
the three most common distance measures between groups:
single, complete, and average linkage.
}
\examples{

# Hierarchical Clustering
clust <- hclust(as.dist(D), method = "single")
plot(clust, hang = -1, main = "Single linkage",
     xlab = "", sub = "")

}
\keyword{datasets}
