% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{farms}
\alias{farms}
\title{farms data}
\format{
A data frame with 120 rows and 3 columns.
}
\usage{
farms
}
\description{
The size of plants have been measured on 5 different locations with various
Nitrogen content in the soil, within in each of 24 farms. This data set
illustrates spatial pseudo-replication, since the 5 measurements made on
each farm are probably dependent, coming from the same farm.
}
\details{
The same data set as in The R Book, but with better column names.
}
\keyword{datasets}
