% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{farms0}
\alias{farms0}
\title{farms0 data}
\format{
A data frame with 80 rows and 3 columns.
}
\usage{
farms0
}
\description{
This is just a subsample of the \code{\link{farms}} data set, where
there are differing number of data from each farm.
}
\keyword{datasets}
