% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{grades}
\alias{grades}
\title{grades data}
\format{
A data frame, 395 observations on 3 variables
}
\usage{
grades
}
\description{
A data set with grades on three subjects, scored from 0 (worst)
to 20 (best)
}
\keyword{datasets}
