% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Hald.Cement}
\alias{Hald.Cement}
\title{Hald's cement data}
\format{
A data frame
}
\usage{
Hald.Cement
}
\description{
A dataset with 13 observations and 5 variables
}
\examples{
summary(Hald.Cement)
}
\keyword{datasets}
