% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hprice.R
\docType{data}
\name{Hprice}
\alias{Hprice}
\title{Hprice: Housing prices and Air Pollution in Boston, USA}
\format{
A data frame with 506 observations (rows) and 12 variables (columns).

\tabular{llllr}{
\tab \bold{Column name} \tab \bold{Data type} \tab \bold{Description} \tab \bold{Values}\cr
\code{[,1]} \tab \code{price} \tab numeric \tab Median housing price in US dollars \tab (5000 - 50001)\cr
\code{[,2]} \tab \code{crime} \tab numeric \tab Crimes committed per capita \tab (0.006 - 88.976)\cr
\code{[,3]} \tab \code{nox} \tab numeric \tab Nitrous oxide in parts per 100 million \tab (3.85 - 8.71)\cr
\code{[,4]} \tab \code{rooms} \tab numeric \tab Average number of rooms per house \tab (3.56 - 8.78)\cr
\code{[,5]} \tab \code{dist} \tab numeric \tab Weighted distance to 5 employ centers in miles \tab (1.13 - 12.13)\cr
\code{[,6]} \tab \code{radial} \tab numeric \tab Accessibility index to radial highways \tab (1 - 24)\cr
\code{[,7]} \tab \code{proptax} \tab numeric \tab Property tax per $1000 \tab (18.7 - 71.1)\cr
\code{[,8]} \tab \code{stratio} \tab numeric \tab Average student-teacher ratio \tab (12.6 - 22.0)\cr
\code{[,9]} \tab \code{lowstat} \tab numeric \tab Percentage of people with 'lower status' \tab (1.73 - 39.07)\cr
\code{[,10]} \tab \code{lprice} \tab numeric \tab log(price) \tab (8.517193 - 10.8198)\cr
\code{[,11]} \tab \code{lnox} \tab numeric \tab log(nox) \tab (1.348073 - 2.164472)\cr
\code{[,12]} \tab \code{lproptax} \tab numeric \tab log(proptax) \tab (5.231109 - 6.566672)
}
}
\source{
Wooldridge, J.M. (2019) \emph{Introductory Econometrics: A Modern Approach,
7th Edition}. Cengage Learning.
}
\usage{
Hprice
}
\description{
This data set is a selection of variables from the Boston house-price
data originally published by Harrison, D. and Rubinfeld, D.L. The data
contains the median housing price in 506 communities in the Boston area,
measurements of air pollution and other information collected by the U.S
Census Service.
}
\examples{

# The structure of the object
str(Hprice)

# Scatterplot matrix with some of the variables
pairs(Hprice[, 1:9])

}
\references{
Harrison, D., Rubinfeld, D.L. (1978) Hedonic prices and the demand for clean
air \emph{J. Environ. Economics & Management}, \bold{5}, 81-102.
}
\keyword{datasets}
