% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LambDataNested.R
\docType{data}
\name{LambDataNested}
\alias{LambDataNested}
\title{LambDataNested: The Weight of Lamb Carcasses}
\format{
A data frame with 240 observations (rows) and 3 variables (columns).
\tabular{llllr}{
\tab \bold{Column name} \tab \bold{Data type} \tab \bold{Description} \tab \bold{Values}\cr
\code{[,1]} \tab \code{Farm} \tab character \tab Lambs from 30 different farms \tab ("Farm1"..."Farm30")\cr
\code{[,2]} \tab \code{Breed} \tab character \tab 5 different sheep breeds \tab ("Kvit spaelsau"..."Gammalnorsk spaelsau")\cr
\code{[,3]} \tab \code{Carcase_weight} \tab numeric \tab Carcass weight in kg \tab (8.2 - 30.4)
}
}
\source{
This is real data from
\href{https://www.animalia.no/no/Dyr/husdyrkontrollene/sauekontrollen/}{"Sauekontrollen"},
a database administered by Animalia AS. The data is provided by Lars Erik
Gangsei.
}
\usage{
LambDataNested
}
\description{
This is real data of the weight of 240 lamb carcasses measured at a
slaughterhouse. The lamb carcasses are from a total of 30 farms, and from
each farm eight lambs were observed. The sheep breed is registered for each
lamb.
}
\examples{

# The structure of the object
str(LambDataNested)

# Encode Farm as factor and check the levels
LambDataNested$Farm <- factor(LambDataNested$Farm,
                              levels = unique(LambDataNested$Farm))
unique(LambDataNested$Farm)

}
\keyword{datasets}
