% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LifeExp.R
\docType{data}
\name{LifeExp}
\alias{LifeExp}
\title{LifeExp: Health Observations in 21 Countries}
\format{
A data frame with 21 observations (rows) and 9 variables (columns).

\tabular{llllr}{
\tab \bold{Column name} \tab \bold{Data type} \tab \bold{Description} \tab \bold{Values}\cr
\code{[,1]} \tab \code{country} \tab factor \tab 21 countries from 5 continents \tab (Argentina...Zambia)\cr
\code{[,2]} \tab \code{pop} \tab integer \tab Population in thousands in 2015 \tab (179 - 229965)\cr
\code{[,3]} \tab \code{growth} \tab numeric \tab The annual growth rate in \% (2005-2015) \tab (-0.6 - 8.7)\cr
\code{[,4]} \tab \code{fert06} \tab numeric \tab The total fertility rate in 2006 \tab (1.2 - 7.5)\cr
\code{[,5]} \tab \code{fert15} \tab numeric \tab The total fertility rate in 2015 \tab (1.4 - 7.1)\cr
\code{[,6]} \tab \code{lifeexp} \tab integer \tab The life expectancy in years \tab (48 - 82)\cr
\code{[,7]} \tab \code{health} \tab integer \tab Per capita total expenditure on health \tab (21 - 6988)\cr
\code{[,8]} \tab \code{water} \tab integer \tab Percent using improved drinking-water sources \tab (47 - 100)\cr
\code{[,9]} \tab \code{alcohol} \tab numeric \tab Alcohol consumption \tab (0.05 - 16.2)
}

\itemize{
\item The per capita total expenditure on health, \code{health}, is the
average exchange rate (US$) in 2014.
\item \code{water} is the percentage of the population that used improved
drinking-water sources in 2014.
\item The data in \code{alcohol} is from 2011 and gives the alcohol
consumption in litres per person among adults aged 15 years.
}
}
\usage{
LifeExp
}
\description{
The life expectancy reported in 2015 and observations related to health were
collected for approximately 200 countries. The 21 countries in this data set
were randomly selected and are from Africa, Europe, Asia, South America, and
Oceania.
}
\examples{

# A short summary of the variables
summary(LifeExp)

# Scatterplot matrix with all variables
pairs(LifeExp)

}
\keyword{datasets}
