% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ListeriaStore.R
\docType{data}
\name{ListeriaStore}
\alias{ListeriaStore}
\title{ListeriaStore: Listeria in Ham}
\format{
A data frame with 48 observations (rows) and 3 variables (columns).
\tabular{llllr}{
\tab \bold{Column name} \tab \bold{Data type} \tab \bold{Description} \tab \bold{Values}\cr
\code{[,1]} \tab \code{ListeriaNumber} \tab numeric \tab Listeria density in cfu/g \tab (635.3076 - 15994102)\cr
\code{[,2]} \tab \code{HamTopping} \tab factor \tab 6 brands of ham \tab (HamT1...HamT6)\cr
\code{[,3]} \tab \code{GroceryStore} \tab factor \tab 4 different stores \tab (Store1...Store4)
}
}
\source{
The data is created by Lars Erik Gangsei.
}
\usage{
ListeriaStore
}
\description{
This is simulated data based on real experiments by Animalia AS (to mimic
realistic variance). The density of the bacterial genus Listeria was measured
on six brands of ham from different grocery stores. The measurements were
done on cold cuts of ham and by counting the number of colony-forming units
per gram.
}
\details{
Cold cuts of ham are available in a large number of brands, differing
slightly in thickness, packing method, producers, addings etc. A research
group wanted to test the hypothesis if the brand of cold cuts of ham had any
effect on Listeria growth. They soon realized that they could not test all
brands. Thus, they randomly picked 6 brands of cold cuts of ham and for those
6 brands they randomly picked 8 replicates. These replicates were picked in 4
different grocery stores, 2 replicates for each brand in each store.

This data was used for compulsory assignment in STAT210 H20, and in Exercises
for STAT210 (exercise 20) in August 2022.
}
\examples{

# First six rows
head(ListeriaStore)

# Linear model
lm(ListeriaNumber ~ HamTopping, data = ListeriaStore)

}
\keyword{datasets}
