% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{litter}
\alias{litter}
\title{litter data}
\format{
A data frame
}
\usage{
litter
}
\description{
60 observations, 3 variables
}
\examples{
library(lme4)
data(litter)
mod <-lmer(Size ~ 1+ (1|Boar) + (1|Farm) + (1|Boar:Farm), data =litter)
summary(mod)
}
\keyword{datasets}
