% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NSRdata}
\alias{NSRdata}
\title{NSRdata data}
\format{
The Norwegian Centre for Science Recruitment (NSR) has an online “education test”
where youths may answer a questionnaire to check their so-called cognitive types, their
science interest, their preferred learning methods and their interest to various science
subjects. The test suggests different ares within the STEM (Science, Technology, Engineering
and Mathematics) within which the youth may find suitable work. We have an excerpt of these
data. The data.frame NSRdata contains two variables, Science and Age. Science is an average
liking score (scale 1-6) to various STEM-subjects, and Age is a factor indicating different
age-groups:
\describe{
\item{a)1: 1-12 yrs}
\item{b)13: 13-15 yrs}
\item{c)16: 16-19 yrs}
\item{d)19: 19-29 yrs}
\item{e)30: 30 + yrs}
}
}
\usage{
NSRdata
}
\description{
Data from the NSR education test
}
\examples{
summary(NSRdata)
}
\keyword{datasets}
