% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Orthodont.unstacked}
\alias{Orthodont.unstacked}
\title{Orthodont.unstacked data}
\format{
A data frame
}
\usage{
Orthodont.unstacked
}
\description{
27observations, 5 variables from the iris data
}
\keyword{datasets}
