% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Pear2011}
\alias{Pear2011}
\title{Pear2011 data, Exam Stat 340 2017-05-19, Exer 1}
\format{
A data frame with
\itemize{
\item \code{Sort} Character.
\item \code{Year} Double. The number deaths from Cardiac Hearth Disease (CHD) per 100000.
\item \code{REF} Double
}
}
\usage{
Pear2011
}
\description{
Data used in Exercise 1, Exam Stat 340 2021-05-28. 24 observations of three
variables.
}
\examples{
# Table 1b, Exam Stat 340 2017-05-19
options(contrasts = c("contr.sum","contr.poly"))
LinearModel.2 <- lm(REF ~ Sort, data = Pear2011)
summary(LinearModel.2)
}
\keyword{datasets}
