% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Penguins.data.R
\docType{data}
\name{Penguins.data}
\alias{Penguins.data}
\title{Penguins.data: Size Measurements of Penguins}
\format{
A data frame with 344 observations (rows) and 8 variables (columns).

\tabular{llllr}{
\tab \bold{Column name} \tab \bold{Data type} \tab \bold{Description} \tab \bold{Values}\cr
\code{[,1]} \tab \code{species} \tab factor \tab 3 different penguin species \tab (Adelie, Chinstrap, Gentoo)\cr
\code{[,2]} \tab \code{island} \tab factor \tab 3 different islands \tab (Biscoe, Dream, Torgersen)\cr
\code{[,3]} \tab \code{bill_length} \tab numeric \tab Culmen length in mm \tab (32.10 - 59.60)\cr
\code{[,4]} \tab \code{bill_depth} \tab numeric \tab Bill depth in mm \tab (13.10 - 21.50)\cr
\code{[,5]} \tab \code{flipper_length} \tab integer \tab Flipper length in mm \tab (172.0 - 231.0)\cr
\code{[,6]} \tab \code{body_mass} \tab integer \tab The penguin's body mass in gram \tab (2700 - 6300)\cr
\code{[,7]} \tab \code{sex} \tab factor \tab The gender of the penguin \tab (female, male)\cr
\code{[,8]} \tab \code{year} \tab integer \tab The year of the study \tab (2007 - 2009)
}

The \code{bill_length} is a measurement of the culmen length: the penguin's
upper ridge of the bill (beak).
}
\source{
Horst, A.M., Hill, A.P., Gorman, K.B. (2020) \pkg{palmerpenguins}: Palmer
Archipelago (Antarctica) penguin data. R package version 0.1.0.
\url{https://allisonhorst.github.io/palmerpenguins/}.
doi: 10.5281/zenodo.3960218.
}
\usage{
Penguins.data
}
\description{
This data is from a study of three penguin species from three islands in
the Palmer Archipelago, Antarctica. Four measurements were collected for 344
male and female penguins in the years 2007 - 2009.
}
\details{
The data is from the \pkg{palmerpenguins} data package which also
contains a data set \code{penguins_raw} with additional information from the
study.
}
\examples{

# A short summary of the variables
summary(Penguins.data)

# Scatterplot matrix of the variables
pairs(Penguins.data)

# Measured flipper length per species
boxplot(flipper_length ~ species, data = Penguins.data,
        col = c("mistyrose4", "snow2", "darkorchid"))

# Plotting flipper length against bill length coloured by species
plot(flipper_length ~ bill_length, data = Penguins.data,
     col = species, pch = 19)

}
\references{
Gorman, K.B., Williams, T.D., Fraser, W.R. (2014) Ecological sexual dimorphism and
environmental variability within a community of Antarctic penguins (genus
Pygoscelis). \emph{PLoS ONE}, \bold{9}, e90081.
\url{https://doi.org/10.1371/journal.pone.0090081}
}
\keyword{datasets}
