% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pines.R
\docType{data}
\name{pines}
\alias{pines}
\title{pines: Experiment with pine trees in growth chambers}
\format{
A data frame with 240 observations (rows) and 4 variables (columns).
\tabular{llllr}{
\tab \bold{Column name} \tab \bold{Data type} \tab \bold{Description} \tab \bold{Values}\cr
\code{[,1]} \tab \code{ddCt} \tab numeric \tab Gene expression values \tab (-17.93 - 8.84)\cr
\code{[,2]} \tab \code{Treatment} \tab integer \tab The foliage treatment \tab (Control, MeJA, PipA)\cr
\code{[,3]} \tab \code{Family} \tab integer \tab The genetic families \tab (a,b,c,d)\cr
\code{[,4]} \tab \code{Box} \tab integer \tab The box where the seeds grow, nested under each Treatment/Family combination \tab (1 - 4)\cr
}
}
\usage{
pines
}
\description{
An experiment where the effect of genetics and some foliage treatments on the
gene expression of pine trees were investigated. There are 4 genetic families
3 treatments in a full factorial design. Then, for each treatment/family
combination seeds were sown into 4 boxes, in total 3x4x4 = 48 boxes. These
were placed randomly in a growth chamber. After some weeks 5 plants from
each box were harvested and gene expression was measured by qPCR.
}
\details{
These data are not from the real experiment, but from a simulation done
before the experiments were conducted, in order to investigate the power
of various analyses under various assumptions
}
\examples{

# Plot
library(ggplot2)
ggplot(pines, aes(x = Treatment, y = ddCt, color = factor(Box))) +
 geom_point(size = 4, alpha = 0.5) +
 facet_wrap(~Family, nrow = 2)

}
\keyword{datasets}
