% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDA.R
\name{plotDA}
\alias{plotDA}
\title{Plotting discriminant analysis regions}
\usage{
plotDA(DAobject = NULL, regions = TRUE, contours = FALSE, resolution = 100)
}
\arguments{
\item{DAobject}{An object from LDA}

\item{regions}{Logical to indicate if regions should be displayed}

\item{contours}{Logical to indicate if contour lines should be displayed.}

\item{resolution}{Integer}
}
\description{
Plot the region in an LDA object.
}
\details{
The LDA object used as (mandatory) input to this function MUST have been made
from data in a \code{data.frame}, and NOT data in a \code{tibble}.
}
