% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{poems.test}
\alias{poems.test}
\title{poems.test data}
\format{
A data frame, 22 observations on 31 variables
}
\usage{
poems.test
}
\description{
The data frame poems contains letter frequencies collected from 22 poems written by either
William Shakespeare (1564–1616), William Blake (1757–1827), and Thomas Stearns Eliot (1888–1965).
}
\keyword{datasets}
