% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pokemon.R
\docType{data}
\name{Pokemon}
\alias{Pokemon}
\title{Pokemon: Pokémon with Stats}
\format{
A data frame with 721 observations (rows) and 12 variables (columns).

\tabular{llllr}{
\tab \bold{Column name} \tab \bold{Data type} \tab \bold{Description} \tab \bold{Values}\cr
\code{[,1]} \tab \code{number} \tab factor \tab Each Pokémon is assigned a unique ID \tab (1... 721)\cr
\code{[,2]} \tab \code{name} \tab character \tab The name of the Pokémon \tab ("Bulbasaur"..."Volcanion")\cr
\code{[,3]} \tab \code{type1} \tab factor \tab 18 different main types \tab (Bug...Water)\cr
\code{[,4]} \tab \code{type2} \tab factor \tab 19 different additional types \tab (Bug...Water)\cr
\code{[,5]} \tab \code{hp} \tab integer \tab Hit points describing the Pokémon's health \tab (1 - 255)\cr
\code{[,6]} \tab \code{attack} \tab integer \tab Attach strength for normal attacks \tab (5 - 165)\cr
\code{[,7]} \tab \code{defense} \tab integer \tab Defence strength against normal attacks \tab (5 - 230)\cr
\code{[,8]} \tab \code{sp.atk} \tab integer \tab Strength for special attack \tab (10 - 154)\cr
\code{[,9]} \tab \code{sp.def} \tab integer \tab Damage resistance against special attacks \tab (20 - 230)\cr
\code{[,10]} \tab \code{speed} \tab integer \tab Attack speed \tab (5 - 160)\cr
\code{[,11]} \tab \code{generation} \tab factor \tab 6 generations of Pokémons \tab (1...6)\cr
\code{[,12]} \tab \code{legendary} \tab numeric \tab 1 if the Pokémon is legendary \tab (0, 1)
}

The type columns, \code{type1} and \code{type2}, refer to different elemental
properties associated with both the Pokémon and the moves (attack).
\itemize{
\item \code{type1}: Each Pokémon has a type, this determines
weakness/resistance to attacks
\item \code{type2}: Some of the Pokémons are dual type and have an
additional type
}

For all the stats variables (\code{hp}, \code{attack}, \code{defense},
\code{sp.atk}, \code{sp.def}, \code{speed}) high values are better than low
values. The hit points, \code{hp}, defines how much damage the Pokémon can
receive before fainting.
}
\source{
The data is from
\href{https://www.kaggle.com/datasets/abcsds/pokemon}{kaggle}.
}
\usage{
Pokemon
}
\description{
This is a data set containing information of 721 types of Pokémon, fictional
creatures from a video game series. For each Pokémon their stats, the type of
the Pokémon, and if it is categorized as legendary or not are registered.
}
\details{
This data was used in the Exam STAT200, May 2022.
}
\examples{

# A short summary of the variables
summary(Pokemon)

# Scatterplot matrix of some of the variables
pairs(Pokemon[, c( "hp", "attack", "defense", "speed")])

}
\keyword{datasets}
