% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{quasibin}
\alias{quasibin}
\title{quasibin data}
\format{
A data frame
\itemize{
\item \code{Success} binary 0-1 response
\item \code{x1} continuous response
\item \code{x2} continuous response
\item \code{x3} continuous response
}
}
\usage{
quasibin
}
\description{
38 observations and 4 five variables (renamed since data is not open)
}
\examples{
# First, look at data
table(quasibin$Success)
boxplot(quasibin$x3 ~ quasibin$Success)

# Logistic regression
logit1 <- glm(Success ~ x3, data = quasibin, family = "binomial")
summary(logit1)

# Suprisingly, `x3` is not a significant predictor.
# Probably reason, observation 7 is an outlier as seen from
plot(logit1, which = 1)

# We model dispersion, more reasonable results:
logit2 <- glm(Success ~ x3, data = quasibin, family = "quasibinomial")
summary(logit2)
}
\keyword{datasets}
