% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Raman.train.R
\docType{data}
\name{Raman.train}
\alias{Raman.train}
\title{Raman.train: Raman Spectroscopy on Fish Oil}
\format{
A data frame with 30 observations (rows) and 2264 variables
(columns).
\tabular{llllr}{
\tab \bold{Column name} \tab \bold{Data type} \tab \bold{Description} \tab \bold{Values}\cr
\code{[,1]} \tab \code{Iodine_number} \tab numeric \tab Indicator for unsaturation in the fatty acids \tab (147.8026 - 169.9538)\cr
\code{[,2:2264]} \tab \code{RSdata1...RSdata2263} \tab numeric \tab Raman spectroscopy data \tab (-0.00348022 - 0.1074767)
}
}
\usage{
Raman.train
}
\description{
This data set contains Raman spectra for 30 salmon oil samples extracted from
farmed salmon, \emph{Salmo salar}. The Iodine number, which is connected to
the degree of unsaturation in the fatty acids, is reported.
}
\details{
The wavelengths used in the Raman spectroscopy were -20 cm\eqn{^{-1}} – 3450
cm\eqn{^{-1}}. Wavelengths before 790 cm\eqn{^{-1}} and after 3050
cm\eqn{^{-1}} are not included in this data.

The Raman data with a total of 45 observations was split into this training
data with 30 observations and a test data set with the remaining 15
observations.
}
\examples{

# Dimensions of the object
dim(Raman.train)

# A short summary of some of the variables
summary(Raman.train[, 1:6])

}
\references{
Liland, K.H., Almøy, T., Mevik, B., et al. (2010) Optimal Choice of Baseline
Correction for Multivariate Calibration of Spectra.
\emph{History of Science}, \bold{64}, 245 - 278.
}
\seealso{
\code{Raman.test}
}
\keyword{datasets}
