% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rootGrowth}
\alias{rootGrowth}
\title{rootGrowth data}
\format{
A data frame with 180 rows and 4 columns.
}
\usage{
rootGrowth
}
\description{
The length of roots is measured bi-weekly for 10 weeks. There
are 6 repeated measurements made for each root. 15 roots grow
under normal conditions (control) while 15 under fertilized
conditions (fertilizer).
}
\details{
This is an extension of the data set fertilizer
in The R Book.
}
\keyword{datasets}
