% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SalmonLice.R
\docType{data}
\name{SalmonLice}
\alias{SalmonLice}
\title{SalmonLice: Salmon Lice and Cleaner Fish}
\format{
A data frame with 48 observations (rows) and 3 variables (columns).
\tabular{llllr}{
\tab \bold{Column name} \tab \bold{Data type} \tab \bold{Description} \tab \bold{Values}\cr
\code{[,1]} \tab \code{SalmonLice} \tab numeric \tab Average number of lice in each cage \tab (0.1210 - 2.9538)\cr
\code{[,2]} \tab \code{Farm} \tab factor \tab 2 levels of farm \tab (Farm 1, Farm 2)\cr
\code{[,3]} \tab \code{CleanerFish} \tab factor \tab 2 types of cleaner fish \tab (Control, Lumpfish, Wrappers)
}

\strong{Note that:} 'Farm1' under 'Lumpfish' is not the same farm as 'Farm1'
under 'Wrappers' and 'Farm1' under control etc. Thus 6 different farms are
represented.
}
\source{
The data is created by Lars Erik Gangsei.
}
\usage{
SalmonLice
}
\description{
This is simulated data. The density of salmon lice is measured as the average
number of adult lice per fish in a fish cage (merd). There are six different
farms and observations for eight fish cages in each of these farms. Lumpfish
(rognkjeks in Norwegian) was used as cleaner fish in two farms and wrasses
(leppefisk in Norwegian) was used in two farms. The remaining two farms was
used as controls (i.e., no cleaner fish).
}
\details{
The parasite salmon lice is a large problem for salmon breeders. It has
become a practice to use a biological method, so-called cleaner fish, to
control the density of salmon lice in the farms. Cleaner fish are released
into the cage along with, or a little time after, salmon is released into the
cage. The number of lice was measured at a fixed time after releasing salmon
into the cage.

This data was used for Exercise 3 in the Exam STAT210 (2019), and in
Exercises for STAT210 (exercise 21) in August 2022.
}
\examples{

# A short summary of the variables
summary(SalmonLice)

# Linear model
lm(SalmonLice ~ CleanerFish, data = SalmonLice)

}
\keyword{datasets}
