% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sediments}
\alias{sediments}
\title{sediments data}
\format{
A data frame with 12 rows and 4113 columns.
}
\usage{
sediments
}
\description{
Seafloor sediments from 12 different sites along the coast have been collected and
their prokaryotic composition estimated by amplicon sequencing and subsequent
bioinformatic analysis.
}
\details{
The column Environment indicates if the samples are from Natural or Polluted sediments.

All remaining columns are the relative abundance of various prokaryotic genera.
}
\keyword{datasets}
