% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sireunstacked.R
\docType{data}
\name{sireunstacked}
\alias{sireunstacked}
\title{sireunstacked: Milk Production from Dairy Cows}
\format{
A data frame with 3 observations (rows) and 3 variables (columns).
\tabular{llllr}{
\tab \bold{Column name} \tab \bold{Data type} \tab \bold{Description} \tab \bold{Values}\cr
\code{[,1]} \tab \code{s1} \tab integer \tab Milk produced from each cow in kg \tab (6789 - 7543)\cr
\code{[,2]} \tab \code{s2} \tab integer \tab Milk produced from each cow in kg \tab (6347 - 6875)
\code{[,3]} \tab \code{s3} \tab integer \tab Milk produced from each cow in kg \tab (6781 - 7435)
}
}
\usage{
sireunstacked
}
\description{
A tiny data set in a "bad" format. This is milk production for
the three daughters (cows) of 3 sires (bulls) used to illustrate
random effect models.
}
\details{
The sireunstacked data set has been used in STAT340 for introducing random effects models. See
\code{\link{SireHerd}} and \code{\link{sires}} for more realistic data sets.
}
\examples{

# Stacking and plotting
sirestacked <- stack(sireunstacked)
boxplot(values ~ ind, data = sirestacked)

}
\seealso{
\code{SireHerd}, \code{sires}.
}
\keyword{datasets}
