% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{small}
\alias{small}
\title{small data}
\format{
\itemize{
\item \code{y}: Binary numeric response, 0 or 1.
\item \code{x}: Predictor variable, 0 or 1.
}
}
\usage{
small
}
\description{
A very small data set used to illustrate logistic regression.
}
\examples{

res1 <- glm(y ~ x, data = small, family = binomial)
coef(res1)

}
\keyword{datasets}
