% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{smallProp}
\alias{smallProp}
\title{small data, accumulated format}
\format{
\itemize{
\item \code{y}: Binary numeric response, 0 or 1.
\item \code{x}: Predictor variable, 0 or 1.
}
}
\usage{
smallProp
}
\description{
A very small data set used to illustrate logistic regression. This is dataset is on an accumulated format, which is suitable for large datasets.
}
\examples{

prop <- smallProp[,1:2]; x <- smallProp[,3]
res1 <- glm(prop ~ x, family = binomial)
coef(res1)

}
\keyword{datasets}
