% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{smoke}
\alias{smoke}
\title{smoke data, Exam Stat 340 2021-05-28, Exer 1}
\format{
A data frame with
\itemize{
\item \code{x} Double. Average number of cigarettes per person per year.
\item \code{y} Double. The number deaths from Cardiac Hearth Disease (CHD) per 100000.
\item \code{country} Character
}
}
\usage{
smoke
}
\description{
Data used in Exercise 1, Exam Stat 340 2018-05-28. 20 observations of three
variables.
}
\examples{
lm(y ~ x, data = smoke)
}
\keyword{datasets}
