% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{splityield}
\alias{splityield}
\title{yields data in R book, Ch 19.4, called \code{splityield} here}
\format{
A data frame
}
\usage{
splityield
}
\description{
72 observations, 5 variables
}
\examples{
library(nlme)
data(splityield)
model <- lme(yield ~ irrigation*density*fertilizer, random = ~1|block/irrigation/density, data = splityield)
summary(model)
}
\keyword{datasets}
