% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STAT210example5.1.R
\docType{data}
\name{STAT210example5.1}
\alias{STAT210example5.1}
\title{STAT210example5.1: The Battery Design Experiment}
\format{
A data frame with 36 observations (rows) and 3 variables (columns).
\tabular{llllr}{
\tab \bold{Column name} \tab \bold{Data type} \tab \bold{Description} \tab \bold{Values}\cr
\code{[,1]} \tab \code{Material_Type} \tab factor \tab 3 types of material plates \tab (1, 2, 3)\cr
\code{[,2]} \tab \code{Temperature} \tab factor \tab 3 levels of temperature in °F \tab (15, 70, 125)\cr
\code{[,3]} \tab \code{Effective_life} \tab integer \tab Battery life in hours \tab (20 - 188)
}
}
\source{
Montgomery, D. C. (2019) \emph{Design and Analysis of Experiments,
9th Edition, EMEA Edition}. New York: Wiley.
}
\usage{
STAT210example5.1
}
\description{
An engineer is designing a battery for use in a device that will be subjected
to some extreme variations in temperature. The engineer tests battery designs
with three different plate materials at three levels of temperatures to
investigate how these parameters affects the effective battery life.
}
\details{
The data is from Example 5.1 in Design and Analysis of Experiments.
9th Edition, EMEA Edition.
}
\examples{

# A short summary of the variables
summary(STAT210example5.1)

# First six and last 10 rows
head(STAT210example5.1)
tail(STAT210example5.1, n = 10)

}
\keyword{datasets}
