% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STAT210example5.2.R
\docType{data}
\name{STAT210example5.2}
\alias{STAT210example5.2}
\title{STAT210example5.2: Impurity in Chemical Products}
\format{
A data frame with 15 observations (rows) and 3 variables (columns).
\tabular{llllr}{
\tab \bold{Column name} \tab \bold{Data type} \tab \bold{Description} \tab \bold{Values}\cr
\code{[,1]} \tab \code{Temperature} \tab factor \tab 3 levels of temperature in °F \tab (100, 125, 150)\cr
\code{[,2]} \tab \code{Pressure} \tab factor \tab 5 levels of pressure \tab (25, 30, 35, 40, 45)\cr
\code{[,3]} \tab \code{Impurity} \tab integer \tab The impurity of each product \tab (1 - 6)
}
}
\source{
Montgomery, D. C. (2019) \emph{Design and Analysis of Experiments,
9th Edition, EMEA Edition}. New York: Wiley.
}
\usage{
STAT210example5.2
}
\description{
The impurity present in a chemical product is affected by two factors:
pressure and temperature. This data shows a factorial experiment with a
single replicate.
}
\details{
The data is from Example 5.2 in Design and Analysis of Experiments,
9th Edition, EMEA Edition.
}
\examples{

# A short summary of the variables
summary(STAT210example5.2)

# Column names
colnames(STAT210example5.2)

# The structure of the object
str(STAT210example5.2)

}
\keyword{datasets}
