% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STAT210prob14.13.R
\docType{data}
\name{STAT210prob14.13}
\alias{STAT210prob14.13}
\title{STAT210prob14.13: Yield Experiment}
\format{
A data frame with 54 observations (rows) and 4 variables (columns).

\tabular{llllr}{
\tab \bold{Column name} \tab \bold{Data type} \tab \bold{Description} \tab \bold{Values}\cr
\code{[,1]} \tab \code{Machine} \tab factor \tab 3 machines were tested \tab (1, 2, 3)\cr
\code{[,2]} \tab \code{Power} \tab factor \tab 2 levels of power settings \tab (1, 2)\cr
\code{[,3]} \tab \code{Station} \tab factor \tab 3 stations per machine  \tab (1, 2, 3)\cr
\code{[,4]} \tab \code{Yield} \tab numeric \tab Product yield from each station per run \tab (22.0 - 37.1)
}
}
\source{
Montgomery, D. C. (2019) \emph{Design and Analysis of Experiments,
9th Edition, EMEA Edition}. New York: Wiley.
}
\usage{
STAT210prob14.13
}
\description{
This is results after a process engineer tested the yield of a product
manufactured on three machines. Each machine can be operated at two power
settings and has three stations on which the product is formed. The machines
were tested at both power settings, and three observations on yield were
taken from each station. The runs were made in random order.
}
\details{
The data is from Exercise 14.13 in Design and Analysis of Experiments,
9th Edition, EMEA Edition. In the exercise all three factors are assumed
fixed.
}
\examples{

# Dimensions of the data frame
dim(STAT210prob14.13)

# Number of replicates
table(STAT210prob14.13[, 1:3])

}
\keyword{datasets}
