% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STAT210prob2.28.R
\docType{data}
\name{STAT210prob2.28}
\alias{STAT210prob2.28}
\title{STAT210prob2.28: Burning Times of Chemical Flares}
\format{
A data frame with 10 observations (rows) and 2 variables (columns).
\tabular{llllr}{
\tab \bold{Column name} \tab \bold{Data type} \tab \bold{Description} \tab \bold{Values}\cr
\code{[,1]} \tab \code{Type_1} \tab integer \tab Burning time in minutes \tab (57 - 82)\cr
\code{[,2]} \tab \code{Type_2} \tab integer \tab Burning time in minutes \tab (56 - 83)
}
}
\source{
Montgomery, D. C. (2019) \emph{Design and Analysis of Experiments,
9th Edition, EMEA Edition}. New York: Wiley.
}
\usage{
STAT210prob2.28
}
\description{
The data in this data set are the burning times of chemical flares of two
different formulations. The design engineers are interested in both the mean
and variance of the burning times.
}
\details{
This is data from Exercise 2.28 in Design and Analysis of Experiments, 9th
Edition, EMEA Edition.
}
\examples{

# The structure of the object
str(STAT210prob2.28)

# A short summary of the variables
summary(STAT210prob2.28)

}
\keyword{datasets}
