% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STAT210prob2.29.R
\docType{data}
\name{STAT210prob2.29}
\alias{STAT210prob2.29}
\title{STAT210prob2.29: The Effect of the Hexafluoroethane Flow Rate}
\format{
A data frame with 12 observations (rows) and 2 variables (columns).
\tabular{llllr}{
\tab \bold{Column name} \tab \bold{Data type} \tab \bold{Description} \tab \bold{Values}\cr
\code{[,1]} \tab \code{FlowRate} \tab integer \tab 2 different C\eqn{_{2}}F\eqn{_{6}} flow rates \tab (125 - 200)\cr
\code{[,2]} \tab \code{Uniformity} \tab numeric \tab The observed uniformity \tab (2.6 - 5.1)
}
}
\source{
Montgomery, D. C. (2019) \emph{Design and Analysis of Experiments,
9th Edition, EMEA Edition}. New York: Wiley.
}
\usage{
STAT210prob2.29
}
\description{
This data is from an article in \emph{Solid State Technology} (May 1987)
which describes an experiment to determine the effect of the
C\eqn{_{2}}F\eqn{_{6}} flow rate on the uniformity of the etch on a silicon
wafer used in integrated circuit manufacturing. All the runs were made in
random order, and the data are the uniformity observations of the two flow
rates.
}
\details{
This is data from Exercise 2.29 in Design and Analysis of Experiments, 9th
Edition, EMEA Edition.
}
\examples{

# A short summary of the variables
summary(STAT210prob2.29)

# Uniformity for different flow rates of Hexafluoroethane
boxplot(Uniformity ~ FlowRate, data = STAT210prob2.29,
        col = "steelblue2")

}
\references{
Yin, G.Z. and Jillie, D.W. (1987) Orthogonal Design for Process
Optimization and Its Application to Plasma Etching.
\emph{Solid State Technology}.
}
\seealso{
\code{STAT210example3.1}, \code{STAT210prob2.31}, \code{STAT210prob2.33}, and \code{STAT210prob2.39}
}
\keyword{datasets}
