% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STAT210prob2.31.R
\docType{data}
\name{STAT210prob2.31}
\alias{STAT210prob2.31}
\title{STAT210prob2.31: Photoresist Thickness for Wafers}
\format{
A data frame with 8 observations (rows) and 2 variables (columns).
\tabular{llllr}{
\tab \bold{Column name} \tab \bold{Data type} \tab \bold{Description} \tab \bold{Values}\cr
\code{[,1]} \tab \code{Thickness_95} \tab numeric \tab Resist thickness for wafers baked at 95°C \tab (6.467 - 11.739)\cr
\code{[,2]} \tab \code{Thickness_100} \tab numeric \tab Resist thickness for wafers baked at 100°C \tab (3.772 - 8.963)
}
}
\source{
Montgomery, D. C. (2019) \emph{Design and Analysis of Experiments,
9th Edition, EMEA Edition}. New York: Wiley.
}
\usage{
STAT210prob2.31
}
\description{
Photoresist is a light-sensitive material applied to semi-conductor wafers so
that the circuit pattern can be imaged on to the wafer. After application,
the coated wafers are baked to remove the solvent in the photoresist mixture
and to harden the resist. This data are the measurements of photoresist
thickness in kA for eight wafers baked at two different temperatures. Assume
that all of the runs were made in random order.
}
\details{
This is data from Exercise 2.31 in Design and Analysis of Experiments, 9th
Edition, EMEA Edition.
}
\examples{

# A short summary of the variables
summary(STAT210prob2.31)

# Stacked data
resist_thickness <- stack(STAT210prob2.31)
colnames(resist_thickness) <- c("thickness", "temperature")

# Ounces per machine
boxplot(thickness ~ temperature, data = resist_thickness,
        col = "mistyrose")

}
\seealso{
\code{STAT210prob2.29}, \code{STAT210example3.1}, \code{STAT210prob2.33}, and \code{STAT210prob2.39}
}
\keyword{datasets}
