% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STAT210prob2.33.R
\docType{data}
\name{STAT210prob2.33}
\alias{STAT210prob2.33}
\title{STAT210prob2.33: Etch Uniformity on Silicon Wafers}
\format{
A data frame with 20 observations (rows) and 1 variable (column).
\tabular{llllr}{
\tab \bold{Column name} \tab \bold{Data type} \tab \bold{Description} \tab \bold{Values}\cr
\code{[,1]} \tab \code{Uniformity} \tab numeric \tab The observed uniformity \tab (4.39 - 7.55)
}
}
\source{
Montgomery, D. C. (2019) \emph{Design and Analysis of Experiments,
9th Edition, EMEA Edition}. New York: Wiley.
}
\usage{
STAT210prob2.33
}
\description{
This data consists of 20 observations on etch uniformity on silicon wafers
taken during a qualification experiment for a plasma etcher.
}
\details{
This is data from Exercise 2.33 in Design and Analysis of Experiments, 9th
Edition, EMEA Edition.
}
\examples{

# Structure of the object
str(STAT210prob2.33)

# A short summary of the variables
summary(STAT210prob2.33)

}
\seealso{
\code{STAT210prob2.29}, \code{STAT210example3.1}, \code{STAT210prob2.31}, and \code{STAT210prob2.39}
}
\keyword{datasets}
