% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STAT210prob2.34.R
\docType{data}
\name{STAT210prob2.34}
\alias{STAT210prob2.34}
\title{STAT210prob2.34: Measurements with Different Calipers}
\format{
A data frame with 12 observations (rows) and 2 variables (columns).
\tabular{llllr}{
\tab \bold{Column name} \tab \bold{Data type} \tab \bold{Description} \tab \bold{Values}\cr
\code{[,1]} \tab \code{Caliper_1} \tab numeric \tab The diameter in inches \tab (0.265 - 0.268)\cr
\code{[,2]} \tab \code{Caliper_2} \tab numeric \tab The diameter in inches \tab (0.264 - 0.269)
}
}
\source{
Montgomery, D. C. (2019) \emph{Design and Analysis of Experiments,
9th Edition, EMEA Edition}. New York: Wiley.
}
\usage{
STAT210prob2.34
}
\description{
Callipers are measuring tools and one of the quickest and most accurate
methods for dimensional measurements. There are however multiple factors that
can affect the measuring accuracy. The diameter of a ball bearing was
measured by 12 inspectors, each using two different kinds of callipers.
}
\details{
This is data from Exercise 2.34 in Design and Analysis of Experiments, 9th
Edition, EMEA Edition.
}
\examples{

# A short summary of the variables
summary(STAT210prob2.34)

# The dimensions of the data frame
dim(STAT210prob2.34)

}
\keyword{datasets}
