% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STAT210prob2.36.R
\docType{data}
\name{STAT210prob2.36}
\alias{STAT210prob2.36}
\title{STAT210prob2.36: Shear Strength of Steel Girder}
\format{
A data frame with 9 observations (rows) and 3 variables (columns).
\tabular{llllr}{
\tab \bold{Column name} \tab \bold{Data type} \tab \bold{Description} \tab \bold{Values}\cr
\code{[,1]} \tab \code{Karlsruhe} \tab numeric \tab Ratio with the Karlsruhe method \tab (1.151 - 1.559)\cr
\code{[,2]} \tab \code{Lehigh} \tab numeric \tab Ratio with the Lehigh method \tab (0.992 - 1.178)\cr
\code{[,3]} \tab \code{Difference} \tab numeric \tab Difference between the methods \tab (0.125 - 0.507)
}
}
\source{
Montgomery, D. C. (2019) \emph{Design and Analysis of Experiments,
9th Edition, EMEA Edition}. New York: Wiley.
}
\usage{
STAT210prob2.36
}
\description{
An article in the \emph{Journal of Strain Analysis} (vol. 18 no. 2, 1983)
compares several procedures for predicting the shear strength for steel plate
girders. This data shows nine girders in the form of the ratio of predicted
to observed load for two of these procedures, the Karlsruhe and Lehigh
methods.
}
\details{
This is data from Exercise 2.36 in Design and Analysis of Experiments, 9th
Edition, EMEA Edition.
}
\examples{

# A summary of the variables
summary(STAT210prob2.36)

# Obtain the difference between the methods
STAT210prob2.36$Karlsruhe - STAT210prob2.36$Lehigh

}
\keyword{datasets}
