% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STAT210prob2.37.R
\docType{data}
\name{STAT210prob2.37}
\alias{STAT210prob2.37}
\title{STAT210prob2.37: Deflection Temperatures for ABS Plastic Pipes}
\format{
A data frame with 12 observations (rows) and 2 variables (columns).
\tabular{llllr}{
\tab \bold{Column name} \tab \bold{Data type} \tab \bold{Description} \tab \bold{Values}\cr
\code{[,1]} \tab \code{Form_1} \tab integer \tab Deflection temperature for formulation 1 \tab (178 - 210)\cr
\code{[,2]} \tab \code{Form_2} \tab integer \tab Deflection temperature for formulation 2 \tab (176 - 206)
}
}
\source{
Montgomery, D. C. (2019) \emph{Design and Analysis of Experiments,
9th Edition, EMEA Edition}. New York: Wiley.
}
\usage{
STAT210prob2.37
}
\description{
The deflection temperature under load for two different formulations of ABS
plastic pipe is being studied. In this data there are two samples of 12
observations each which are prepared using each formulation. The deflection
temperatures (in °F) are reported.
}
\details{
This is data from Exercise 2.37 in Design and Analysis of Experiments, 9th
Edition, EMEA Edition.
}
\examples{

# Number of rows and columns
nrow(STAT210prob2.37)
ncol(STAT210prob2.37)

# A short summary of the variables
summary(STAT210prob2.37)

}
\keyword{datasets}
