% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STAT210prob2.45.R
\docType{data}
\name{STAT210prob2.45}
\alias{STAT210prob2.45}
\title{STAT210prob2.45: Develop Equation 2.50}
\format{
A data frame with 10 observations (rows) and 3 variables (columns).
\tabular{llllr}{
\tab \bold{Column name} \tab \bold{Data type} \tab \bold{Values}\cr
\code{[,1]} \tab \code{A} \tab numeric \tab (-15.9034 - 19.9977)\cr
\code{[,2]} \tab \code{B} \tab numeric \tab (-15.0013 - 21.1018)\cr
\code{[,3]} \tab \code{delta} \tab numeric \tab (-1.4324 - -0.0965)
}
}
\source{
Montgomery, D. C. (2019) \emph{Design and Analysis of Experiments,
9th Edition, EMEA Edition}. New York: Wiley.
}
\usage{
STAT210prob2.45
}
\description{
This data is for developing the following equation...
\deqn{\frac{{S_1}^2}{{S_2}^2} {F_{1-\alpha/2,{n_2}-1,{n_1}-1}}
\frac{<}{ } \frac{{\sigma_1}^2}{{\sigma_2}^2} \frac{<}{ }
\frac{{S_1}^2}{{S_2}^2} {F_{\alpha/2,{n_2}-1,{n_1}-1}}}
...for a 100(1 - \eqn{\alpha}) percent confidence interval for the ratio
\eqn{{\sigma_1}^2/{\sigma_2}^2}, where \eqn{{\sigma_1}^2} and
\eqn{{\sigma_2}^2} are the variances of two normal distributions.
}
\details{
This is data from Exercise 2.45 in Design and Analysis of Experiments, 9th
Edition, EMEA Edition, and for developing equation 2.50 from the textbook.
}
\examples{

# A short summary of the variables
summary(STAT210prob2.45)

}
\keyword{datasets}
