% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STAT210prob2.46.R
\docType{data}
\name{STAT210prob2.46}
\alias{STAT210prob2.46}
\title{STAT210prob2.46: Equation for CI}
\format{
A data frame with 10 observations (rows) and 2 variables (columns).
\tabular{llllr}{
\tab \bold{Column name} \tab \bold{Data type} \tab \bold{Values}\cr
\code{[,1]} \tab \code{Type_1} \tab integer \tab (57 - 82)\cr
\code{[,2]} \tab \code{Type_2} \tab integer \tab (56 - 83)
}
}
\source{
Montgomery, D. C. (2019) \emph{Design and Analysis of Experiments,
9th Edition, EMEA Edition}. New York: Wiley.
}
\usage{
STAT210prob2.46
}
\description{
This is data for developing an equation for finding a 100(1 - \eqn{\alpha})
percent confidence interval on the difference in the
means of two normal distributions where \eqn{{\sigma_1}^2 != {\sigma_2}^2}.
}
\details{
This is data from Exercise 2.46 in Design and Analysis of Experiments, 9th
Edition, EMEA Edition. In this exercise you are suppose to apply the equation
to the Portland cement experiment data, and find a 95 percent confidence
interval.
}
\examples{

# A short summary of the variables
summary(STAT210prob2.46)

}
\keyword{datasets}
