% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STAT210prob3.19.R
\docType{data}
\name{STAT210prob3.19}
\alias{STAT210prob3.19}
\title{STAT210prob3.19: Donations to the Opera}
\format{
A data frame with 24 observations (rows) and 3 variables (columns).

\tabular{llllr}{
\tab \bold{Column name} \tab \bold{Data type} \tab \bold{Description} \tab \bold{Values}\cr
\code{[,1]} \tab \code{Approach} \tab integer \tab 3 different approaches \tab (1, 2, 3)\cr
\code{[,2]} \tab \code{Contribution} \tab integer \tab The donation from each sponsor in $ \tab (900 - 2000)\cr
\code{[,3]} \tab \code{RESI1} \tab numeric \tab Residuals \tab (-537.5 - 493.75)
}
}
\source{
Montgomery, D. C. (2019) \emph{Design and Analysis of Experiments,
9th Edition, EMEA Edition}. New York: Wiley.
}
\usage{
STAT210prob3.19
}
\description{
A regional opera company has tried three approaches to solicit donations from 24
potential sponsors. The sponsors were randomly divided into three groups of
eight and one approach was used for each group. This data shows the dollar
amounts of the resulting contributions for each of the sponsors.
}
\details{
This is data from Exercise 3.19 in Design and Analysis of Experiments, 9th
Edition, EMEA Edition.
}
\examples{

# A short summary of the variables
summary(STAT210prob3.19)

# Get the residuals
STAT210prob3.19$Approach <- as.factor(STAT210prob3.19$Approach)
res <- aov(Contribution ~ Approach, data = STAT210prob3.19)
res$residuals

}
\keyword{datasets}
