% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stat340_raq_quiz}
\alias{stat340_raq_quiz}
\title{stat340_raq_quiz data}
\format{
During the STAT340 course given in 2019 and 2020, the students were
asked to fill out the ‘R Anxiety Questionnaire’ and a STAT340 quiz. The ‘R
Anxiety Questionnaire’ consists of 24 claims, and the students gave a score
of 1-5 to each claim (a score of 1 implied that they strongly disagreed, 5
implied that they strongly agreed). The STAT340 quiz consists of 20 questions,
and a total of 20 points could be scored.
}
\usage{
stat340_raq_quiz
}
\description{
Data for Group-Exercises-Multivariate II
}
\examples{
summary(stat340_raq_quiz)
}
\keyword{datasets}
