% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{state.x77}
\alias{state.x77}
\title{state.x77 data}
\format{
\describe{
\item{Population:}{population estimate as of July 1, 1975}
\item{Income:}{per capita income (1974)}
\item{Illiteracy:}{illiteracy (1970, percent of population)}
\item{Life Exp:}{life expectancy in years (1969–71)}
\item{Murder:}{murder and non-negligent manslaughter rate per 100,000 population (1976)}
\item{HS Grad:}{percent high-school graduates (1970)}
\item{Frost:s}{mean number of days with minimum temperature below freezing (1931–1960) in capital or large city}
\item{Area:}{land area in square miles}
}
}
\usage{
state.x77
}
\description{
Matrix with 50 rows and 8 columns giving the following statistics in the respective columns.
}
\examples{
summary(state.x77)
}
\keyword{datasets}
