% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{trackfieldrecords}
\alias{trackfieldrecords}
\title{trackfieldrecords}
\format{
A list with two elements, \code{runMen} and \code{runWomen}
}
\usage{
trackfieldrecords
}
\description{
Data for men and women
}
\examples{
data(trackfieldrecords)
clust <-hclust(dist(runWomen[,-8]), method = "average")
plot(clust, hang = -1, xlab = "", sub ="", cex = 0.6, labels = runWomen[,8])
}
\keyword{datasets}
