% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TrackRecords}
\alias{TrackRecords}
\title{TrackRecords data}
\format{
A data frame
}
\usage{
TrackRecords
}
\description{
A dataset with 54 observations and 8 variables, 100m  -marathon recorrds
}
\examples{
summary(TrackRecords)
}
\keyword{datasets}
