% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bears}
\alias{bears}
\title{bears data}
\format{
A data frame with 12 four continuous variables
}
\usage{
bears
}
\description{
A dataset with 24 observations on 12 variables
}
\examples{
summary(bears)
}
\keyword{datasets}
