% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Birds.R
\docType{data}
\name{Birds}
\alias{Birds}
\title{Birds: Winter Bird Species in Norway}
\format{
A data frame with 25 observations (rows) and 3163 variables
(columns).

There is one row for each bird species. The row names consists of the names
of the birds, both in English and Norwegian. The value reflects the largest
simultaneous count of a given bird species in a square region.
}
\usage{
Birds
}
\description{
Birds were observed in 10 km\eqn{^2} squares across entire mainland Norway,
which was divided into 3163 squares in total. Every square is considered one
variable. We consider the 25 most abundant winter bird species in Norway.
}
\examples{

# The bird species in the data
rownames(Birds)

# The dimensions of the data frame
dim(Birds)

}
\seealso{
\code{Birds4} and \code{Birdcount}
}
\keyword{datasets}
