% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Birds4}
\alias{Birds4}
\title{Birds4 data}
\format{
A data frame with 4 rows (one for each bird species) and 3163 columns.
The value reflects the largest simultaneous count
of a given bird species in a square region.
}
\usage{
Birds4
}
\description{
Birds were observed in 10 km2 squares across entire mainland Norway,
which was divided into 3163 squares in total.
Every square is considered one variable. We
consider 4 winter bird species in Norway.
}
\seealso{
\code{Birds}
}
\keyword{datasets}
