% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bodydata}
\alias{bodydata}
\title{bodydata}
\format{
A data frame with 4 four continuous variables
\itemize{
\item \code{Weight} in kg
\item \code{Height} in cm
\item \code{Age} in years
\item \code{Circumference} in cm
}
}
\usage{
bodydata
}
\description{
A dataset with 407 observations on variables
}
\examples{
summary(bodydata)
}
\keyword{datasets}
