% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exer}
\alias{exer}
\title{exer data
The data are from statistics.ats.ucla.edu.
The data called exer, consists of people who were randomly assigned to
two different diets: low-fat and not low-fat and three different
types of exercise: at rest, walking leisurely and running.
Their pulse rate was measured at three different time
points during their assigned exercise: at 1 minute, 15 minutes and 30 minutes.}
\format{
A data frame, 90 observations on 5 variables
}
\usage{
exer
}
\description{
exer data
The data are from statistics.ats.ucla.edu.
The data called exer, consists of people who were randomly assigned to
two different diets: low-fat and not low-fat and three different
types of exercise: at rest, walking leisurely and running.
Their pulse rate was measured at three different time
points during their assigned exercise: at 1 minute, 15 minutes and 30 minutes.
}
\keyword{datasets}
