% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fertilizer}
\alias{fertilizer}
\title{fertilizer data}
\format{
A data frame
}
\usage{
fertilizer
}
\description{
60 observations and 4 variables
}
\examples{
str(fertilizer)
}
\keyword{datasets}
