% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fertilizer}
\alias{fertilizer}
\title{fertilizer data}
\format{
A data frame with variables
\itemize{
\item root: Continuous. Length of \code{plant}.
\item week: Integer. Time of measurement.
\item plant: Factor. ID for plant
\item fertilizer. Factor with values added or control.
}
}
\source{
\verb{The R book}, M.J. Crawley, Wiley, 2012.
}
\usage{
fertilizer
}
\description{
The length of plants as measured by \code{root} is recorded for
60 plants, each plant observed at \code{week} 2, 4, ..., 10.
30 plants have fertilizer added, 30 are controls
}
\examples{
library(nlme)
res <- lme(fixed = root ~ fertilizer,
       random = ~ week | plant,
       data = fertilizer)
}
\keyword{datasets}
