% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{iris.test}
\alias{iris.test}
\title{iris.test data}
\format{
A data frame for testing the classification model from the iris.train dataset
}
\usage{
iris.test
}
\description{
A dataset with 60 observations
}
\examples{
summary(iris.test)
}
\keyword{datasets}
