% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Orthodont.unstacked.R
\docType{data}
\name{Orthodont.unstacked}
\alias{Orthodont.unstacked}
\title{Orthodont.unstacked: Measurements from a Dental Study on Children}
\format{
A data frame with 27 observations (rows) and 5 variables (columns).
\tabular{llllr}{
\tab \bold{Column name} \tab \bold{Data type} \tab \bold{Description} \tab \bold{Values}\cr
\code{[,1]} \tab \code{gender} \tab factor \tab The child's gender \tab (F, M)\cr
\code{[,2]} \tab \code{age8} \tab numeric \tab The distance in mm measured at age 8 \tab (16.5 - 27.5)\cr
\code{[,3]} \tab \code{age10} \tab numeric \tab The distance in mm measured at age 10 \tab (19 - 28)\cr
\code{[,4]} \tab \code{age12} \tab numeric \tab The distance in mm measured at age 12 \tab (19 - 31)\cr
\code{[,5]} \tab \code{age14} \tab numeric \tab The distance in mm measured at age 14 \tab (19.5 - 31.5)
}
}
\usage{
Orthodont.unstacked
}
\description{
The data in this data set is from measurements of 27 children at the ages 8,
10, 12 and 14 (11 girls and 16 boys) and was collected by investigators at
the University of North Carolina Dental School. By the use of x-ray, they
measured the children's distance between the centre of the pituitary and the
pterygomaxillary fissure.
}
\examples{

# Combine the age columns
stacked_orthodont <- cbind(Orthodont.unstacked[1],
                           stack(Orthodont.unstacked[2:5]))

# Rename columns
colnames(stacked_orthodont)[2:3] <- c("Distance", "Age")

# Measured distance based on gender and age
boxplot(Distance ~ gender*Age, data = stacked_orthodont)

}
\references{
Potthoff, R. F. and Roy, S. N. (1964) A Generalized Multivariate Analysis of
Variance Model Useful Especially for Growth Curve Problems.
\emph{Biometrika}, \bold{51}, 313--326.
}
\keyword{datasets}
