% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poems.test.R
\docType{data}
\name{poems.test}
\alias{poems.test}
\title{poems.test: Letter Frequencies in Six Poems}
\format{
A data frame with 6 observations (rows) and 31 variables (columns).

The first column, \code{[,1]} \code{Author}, is a factor and specifies the
surname of the author of the given poem. The remaining columns are all
numeric, and there is one column for each letter in the alphabet (a-z) and
four columns for special characters.
}
\usage{
poems.test
}
\description{
This data set contains letter frequencies collected from six poems written by
either William Shakespeare (1564–1616), William Blake (1757–1827), or Thomas
Stearns Eliot (1888–1965). There are two poems per author in this data.
}
\examples{

# The letters
colnames(poems.test[, -1])

}
\seealso{
\code{poems}
}
\keyword{datasets}
