% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{reading}
\alias{reading}
\title{reading data}
\format{
A data frame with 6 rows and 3 columns.
}
\usage{
reading
}
\description{
The reading ability of children measured at different ages. A tiny
data set to illustrate the importance of dependency in data.
}
\keyword{datasets}
