% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stat210_5Jan2018}
\alias{stat210_5Jan2018}
\title{stat210_5Jan2018 data}
\format{
A data set contains observations of different fertilizer mixtures
coded by the factor variable mixture. The response variable Y called yield
was measured and the year was recorded.
}
\usage{
stat210_5Jan2018
}
\description{
Data for Group-Exercises-ANOVA from the Exam STAT210 (January 5th, 2018)
}
\examples{
summary(stat210_5Jan2018)
}
\keyword{datasets}
