% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{weights}
\alias{weights}
\title{weights data}
\format{
A data frame with two factors and a continuous variable
}
\usage{
weights
}
\description{
A dataset with 48 observations
}
\examples{
summary(weights)
}
\keyword{datasets}
