% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{yields}
\alias{yields}
\title{yields data}
\format{
A data frame with one continuous variable (\code{yield}) and one factor (\code{soil})
}
\usage{
yields
}
\description{
A dataset with 30 observations
}
\examples{
summary(yields)
}
\keyword{datasets}
