% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{yields.unstacked}
\alias{yields.unstacked}
\title{yields unstacked data}
\format{
A data frame with values for \code{yield} for three factor values
}
\usage{
yields.unstacked
}
\description{
A dataset with 10 lines
}
\examples{
summary(yields.unstacked)
}
\keyword{datasets}
